requirements.txt

python-dotenv

# data processing:
pandas
#tables # pytables package required when saving data to hd5 format
numpy

# dataviz:
matplotlib
#seaborn
plotly
kaleido # for exporting plotly images to png / html

# embeddings:
openai==0.28 # ok so there is now a 1.0 interface but we originally obtained using earlier API, so pinning that here


# machine learning:
scikit-learn #==1.3.0
#category_encoders
#tensorflow
#xgboost==1.7.6 #==1.2.1 #==1.5.0 # 1.7.6
#lightgbm
#umap-learn
#kmodes
#hdbscan

# NLP:
#gensim
#python-Levenshtein


# model storage:
google-cloud-storage


# automated tests:
pytest
